backend/src/app/requirements.txt
sqlalchemy
mysqlclient
marshmallow-sqlalchemy
mysql-connector-python